Add: segment cleanup interface.
