+++ src/storage/buffer/buffer_manager.cpp
@@ -74,6 +74,29 @@ BufferObj *BufferManager::Get(UniquePtr<FileWorker> file_worker) {
     return iter2->second.get();
 }
 
+// return false if buffer_obj is not loaded
+bool BufferManager::Cleanup(const String &file_path) {
+    UniquePtr<BufferObj> buffer_obj = nullptr;
+    {
+        std::unique_lock w_lock(rw_locker_);
+        if (auto iter = buffer_map_.find(file_path); iter != buffer_map_.end()) {
+            buffer_obj = std::move(iter->second);
+            buffer_map_.erase(iter);
+        }
+    }
+    if (buffer_obj.get() == nullptr) {
+        // FIXME: is it right?
+        LocalFileSystem fs;
+        if (!fs.Exists(file_path)) {
+            UnrecoverableError(fmt::format("File {} not found.", file_path));
+        }
+        fs.DeleteFile(file_path);
+        return false;
+    }
+    buffer_obj->Cleanup();
+    return true;
+}
+
 void BufferManager::RequestSpace(SizeT need_size, BufferObj *buffer_obj) {
     while (current_memory_size_ + need_size > memory_limit_) {
         BufferObj *buffer_obj1 = nullptr;
