+++ src/storage/io/file_system.cppm
@@ -78,6 +78,8 @@ public:
 
     virtual u64 DeleteDirectory(const String &path) = 0;
 
+    virtual u64 DeleteEmptyDirectory(const String &path) = 0;
+
     virtual Vector<SharedPtr<DirEntry>> ListDirectory(const String &path) = 0;
 
     inline FileSystemType file_system_type() const {
